# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "系统后端 API",
        "title": "Go DDD Seed API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/go-ddd-seed/go-ddd-seed/support",
            "email": "team@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/infra/dict-entries": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据字典类型ID获取所有字典条目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "查询字典条目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典类型ID",
                        "name": "type_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "字典条目列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictEntryResp"
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "向指定字典类型添加新的字典条目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "添加字典条目",
                "parameters": [
                    {
                        "description": "添加字典条目请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.DictEntryAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功，返回包含条目的字典类型",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "字典类型不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/infra/dict-entries/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据ID获取字典条目详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "按ID查询字典条目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "字典条目详情",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictEntryResp"
                        }
                    },
                    "404": {
                        "description": "字典条目不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新指定字典条目的标签、值和排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "更新字典条目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新字典条目请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.DictEntryUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功，返回包含条目的字典类型",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "字典条目不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定的字典条目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "删除字典条目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功，返回更新后的字典类型",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                        }
                    },
                    "404": {
                        "description": "字典条目不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/infra/dict-entries/{id}/disable": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "禁用指定的字典条目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "禁用字典条目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "禁用成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                        }
                    },
                    "404": {
                        "description": "字典条目不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/infra/dict-entries/{id}/enable": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "启用指定的字典条目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "启用字典条目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "启用成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                        }
                    },
                    "404": {
                        "description": "字典条目不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/infra/dict-types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "分页查询字典类型列表，支持按状态过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "字典类型列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "description": "每页数量，默认20",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态过滤：-1=全部 0=禁用 1=启用",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "字典类型列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.PaginatedResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新的字典类型定义",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "创建字典类型",
                "parameters": [
                    {
                        "description": "创建字典类型请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.DictTypeCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/infra/dict-types/code/{code}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据编码获取字典类型详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "基础设施-字典管理"
                ],
                "summary": "按编码查询字典类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典类型编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "字典类型详情",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictTypeResp"
                        }
                    },
                    "404": {
                        "description": "字典类型不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 2
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                }
            }
        },
...
        "github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.DictEntryAddReq": {
            "type": "object",
            "required": [
                "label",
                "type_id",
                "value"
            ],
            "properties": {
                "label": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "sort_order": {
                    "type": "integer",
                    "minimum": 0
                },
                "type_id": {
                    "type": "string"
                },
                "value": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                }
            }
        },
...
        "github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.DictEntryResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication. Format: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}